third_party/blink/web_tests/virtual/controls-refresh/month-picker/month-picker-this-month-button.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../fast/forms/resources/common.js"></script>
<script src="../../../fast/forms/resources/picker-common.js"></script>
<script src="../../../fast/forms/calendar-picker/resources/calendar-picker-common.js"></script>
</head>
<body>
<input type="month" id="month" value="2019-02">
<script>

promise_test(() => {
  let monthElement = document.getElementById('month');
  return openPickerWithPromise(monthElement)
  .then(() => {
    clickThisMonthButton();

    let splitDate = monthElement.value.split('-');
    let actualTodayDateString = new Date(splitDate[0], splitDate[1] - 1).toDateString();
    let today = new Date();
    today.setDate(1);
    let expectedTodayDateString = today.toDateString();

     assert_equals(actualTodayDateString, expectedTodayDateString, "'This month button should have updated in-page control to this month");
     assert_equals(internals.pagePopupWindow, null, "Click on 'This month button should close popup.");
  });
}, "Date picker: 'This month' button should select current month");

</script>
</body>
</html>
